refactor zeitwerk compliance db/seeds/gobierto_seeds/
