+++ db/seeds/gobierto_seeds/gobierto_people/person_statement_seeds.yml
